+++ controllers/FullCalendarController.php
@@ -3,6 +3,7 @@
 namespace humhub\modules\calendar\controllers;
 
 use DateTime;
+use humhub\components\access\StrictAccess;
 use Yii;
 use yii\web\HttpException;
 use humhub\components\Controller;
@@ -12,6 +13,14 @@ use humhub\modules\content\models\Content;
 
 class FullCalendarController extends Controller
 {
+    /**
+     * @inheritdocs
+     */
+    public $access = StrictAccess::class;
+
+    /**
+     * @inheritdocs
+     */
     public function getAccessRules()
     {
         return [